feat: separate out types
