+++ packages/react/package.json
@@ -42,6 +42,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^14.2.2",
     "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^28.1.4",
     "@types/react": ">=16.9.0 <19.0.0",
     "@types/react-dom": ">=16.9.0 <19.0.0",
