+++ packages/react/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './IReactInitSpaceProps'
 export * from './IReactSpaceProps'
 export * from './iFrameProps'
+export * from './IFlatfileProvider'
